+++ frontend/src/components/modal/GuideModal.tsx
@@ -4,7 +4,7 @@ import { Bubble, CloseIcon } from '../common/icons/Icons';
 import { BodyKrRegular3, HeadingKrMedium7 } from '../../styles/typefaces';
 import CenterWrapper from '../layout/CenterWrapper';
 import { DimmedBackground } from './DimmedBackground';
-import { GuideModalContext } from '../../context/GuideMoadlContext';
+import { GuideModalContext } from '../../context/GuideMoadlProvider';
 import { useLocation } from 'react-router-dom';
 import { PATH } from '../../utils/constants';
